nona_back/test/test_BusAPI.py
import unittest
from unittest.mock import MagicMock
from sqlalchemy.orm import sessionmaker
from flask import Flask
from packages.models.db.tb_user import tb_user as User
from packages.models.db.tb_service import tb_service as Service
from packages.models.db.tb_business import tb_business as Business
from packages.models.db.tb_order import OrderStatus, tb_order as Order
from packages.models.db.tb_service_weekly_time_slots import tb_service_weekly_time_slots as ServiceWeeklyTimeSlot
from packages.models.db.tb_service_actual_time_slots import tb_service_actual_time_slots as ServiceActualTimeSlot
from packages.apis.BusAPI import *
from packages.utils.QueryConverter import QueryConverter
import datetime as dt
import json, io


class TestBusAPI(unittest.TestCase):

    def setUp(self):
        self.app = Flask(__name__)
        self.app.config['CORS_HEADERS'] = 'Content-Type'
        self.converter = QueryConverter()
        self.engine = MagicMock()
        self.DBSession = sessionmaker(bind=self.engine)
        self.business_model = BusinessModel(self.engine)
        self.business_model.DBSession = MagicMock()
        self.business_model.session = MagicMock()
        self.mock_session = self.business_model.session.return_value
        self.business_model.session.return_value = self.mock_session

        self.api_key = 'test_api_key'
        self.token ='test_token'

    def test_get_bus_user_from_api_key(self):
        mock_user = User()
        mock_user.username = 'test_username'
        mock_user.email = 'test_email'
        mock_user.uid = 'test_uid'
        mock_user.bus_user = True
        mock_user.bus_api = self.api_key

        self.business_model.get_user_from_api_key = MagicMock(return_value=mock_user)
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get('/api/bus/test_api_key/bus_user')
        self.assertEqual(response.status_code, 200)
        self.business_model.get_user_from_api_key.assert_called_once_with(self.api_key)

    def test_get_bus_user_from_api_key_exception(self):
        error_message = "Test exception"

        self.business_model.get_user_from_api_key = MagicMock(side_effect=Exception(error_message))
        bus_api_register(self.app, self.business_model)

        with self.app.test_client() as client:
            response = client.get('/api/bus/test_api_key/bus_user')

        self.assertEqual(response.status_code, 500)  # Assuming return_error() sets the status code to 500
        self.assertIn(error_message, response.get_data(
            as_text=True))  # Assuming return_error() includes the error message in the response
        self.business_model.get_user_from_api_key.assert_called_once_with(self.api_key)

    def test_get_api_key_from_bus_user(self):
        test_token = 'test_token'
        test_api_key = 'test_api_key'

        self.business_model.get_api_key_from_user = MagicMock(return_value=test_api_key)
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post(f'/api/bus/api_key', json={'token': test_token})
        response.status_code = 200
        self.assertEqual(response.status_code, 200)

    def test_get_api_key_from_bus_user_success(self):
        self.business_model.get_api_key_from_user = MagicMock(return_value=self.api_key)
        bus_api_register(self.app, self.business_model)

        with self.app.test_client() as client:
            response = client.post('/api/bus/api_key',
                                   data=json.dumps({"token": self.token}),
                                   content_type='application/json')

        self.assertEqual(response.status_code, 200)
        self.assertEqual(response.get_data(as_text=True), self.api_key)
        self.business_model.get_api_key_from_user.assert_called_once_with(self.token)

    def test_get_api_key_from_bus_user_exception(self):
        error_message = "Test exception"

        self.business_model.get_api_key_from_user = MagicMock(side_effect=Exception(error_message))
        bus_api_register(self.app, self.business_model)

        with self.app.test_client() as client:
            response = client.post('/api/bus/api_key',
                                   data=json.dumps({"token": self.token}),
                                   content_type='application/json')

        self.assertEqual(response.status_code, 500)  # Assuming return_error() sets the status code to 500
        self.assertIn(error_message, response.get_data(as_text=True))  # Assuming return_error() includes the error message in the response
        self.business_model.get_api_key_from_user.assert_called_once_with(self.token)


    def test_get_services_with_time_slots(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)
        mock_services = [
            (Service(), ServiceWeeklyTimeSlot()),
            (Service(), ServiceWeeklyTimeSlot()),
        ]
        self.business_model.session.return_value.__enter__.return_value.query.return_value.join.return_value.filter.return_value.all.side_effect = [
            mock_services]

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get('/api/bus/test_api_key/services')
        # Assert
        self.assertEqual(response.status_code, 200)
        #self.business_model.get_all_services_with_time_slots.assert_called_once_with(self.api_key)
        self.business_model.session.return_value.__enter__.return_value.query.assert_called()


    # def test_get_services_with_time_slots(self):
    #     mock_user = User()
    #     mock_user.id = 1
    #     self.business_model.get_business_user = MagicMock(return_value=mock_user)
    #     mock_services = [
    #         (Service(), ServiceWeeklyTimeSlot()),
    #         (Service(), ServiceWeeklyTimeSlot()),
    #     ]
    #     self.business_model.get_all_services_with_time_slots = MagicMock(return_value=mock_services)
    #     self.converter.service_with_weekly_time_slots_to_json = MagicMock(return_value="fake_json")
    #
    #     bus_api_register(self.app, self.business_model)
    #     with self.app.test_client() as client:
    #         response = client.get(f'/api/bus/{self.api_key}/services')
    #
    #     self.assertEqual(response.status_code, 200)
    #     self.business_model.get_all_services_with_time_slots.assert_called_once_with(self.api_key)
    #     self.converter.service_with_weekly_time_slots_to_json.assert_called_once_with(mock_services)

    def test_get_all_time_slots(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        service_id = 1
        include_past = False

        mock_time_slots = [
            ServiceActualTimeSlot(id=1, service_id=1, starts=dt.datetime(2023, 4, 7, 9, 0, 0)),
            ServiceActualTimeSlot(id=2, service_id=1, starts=dt.datetime(2023, 4, 7, 10, 0, 0)),
        ]

        self.business_model.get_actual_time_slots = MagicMock(return_value=mock_time_slots)
        self.converter.actual_time_slots_to_json = MagicMock(return_value="fake_json")

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get(f"/api/bus/{self.api_key}/{service_id}/all_time_slots")

        self.assertEqual(response.status_code, 200)
        self.business_model.get_actual_time_slots.assert_called_once_with(self.api_key, str(service_id), include_past)
        #self.converter.actual_time_slots_to_json.assert_called_once_with(mock_time_slots)


    def test_get_available_time_slots(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        service_id = 1

        mock_time_slots = [
            ServiceActualTimeSlot(id=1, service_id=1, starts=dt.datetime(2023, 4, 7, 9, 0, 0)),
            ServiceActualTimeSlot(id=2, service_id=1, starts=dt.datetime(2023, 4, 7, 10, 0, 0)),
        ]

        self.business_model.get_available_time_slots = MagicMock(return_value=mock_time_slots)

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get(f'/api/bus/{self.api_key}/{service_id}/available_time_slots')

        self.assertEqual(response.status_code, 200)
        self.business_model.get_available_time_slots.assert_called_once_with(self.api_key, str(service_id), False)

    def test_get_orders_with_time_slots(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        service_id = 1
        include_past = False

        mock_orders = [
            (Order(), ServiceActualTimeSlot()),
            (Order(), ServiceActualTimeSlot()),
        ]

        self.business_model.get_orders = MagicMock(return_value=mock_orders)
        self.converter.orders_with_time_to_json = MagicMock(return_value="fake_json")

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get(f"/api/bus/{self.api_key}/{service_id}/orders")
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.business_model.get_orders.assert_called_once_with(self.api_key, str(service_id), include_past)
        #self.converter.orders_with_time_to_json.assert_called_once_with(mock_orders)

    def test_get_orders(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        include_past = False

        mock_orders = [
            (Order(), ServiceActualTimeSlot()),
            (Order(), ServiceActualTimeSlot()),
        ]

        self.business_model.get_all_orders = MagicMock(return_value=mock_orders)
        self.converter.orders_with_time_to_json = MagicMock(return_value="fake_json")

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get(f"/api/bus/{self.api_key}/orders")
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.business_model.get_all_orders.assert_called_once_with(self.api_key, include_past)
        #self.converter.orders_with_time_to_json.assert_called_once_with(mock_orders)







    def test_create_bus_user(self):
        test_token = 'test_token'
        test_business_name = 'test_business_name'
        test_admin_api = 'test_admin_api'
        test_create_result = 'test_create_result'

        self.business_model.create_business = MagicMock(return_value=test_create_result)
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.put('/api/admin/bus_user',
                                  json={'token': test_token, 'business_name': test_business_name,
                                        'admin_api': test_admin_api})

        self.assertEqual(response.status_code, 200)
        self.business_model.create_business.assert_called_once_with(test_token, test_admin_api, test_business_name)

    def test_create_service(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        service_data = {
            "api_key": self.api_key,
            "service_name": "test_service",
            "service_description": "test_description",
            "service_price": 100,
            "service_time": [
                {"weekday": "monday", "hour": 9, "minute": 0, "slots": 5},
                {"weekday": "tuesday", "hour": 10, "minute": 0, "slots": 3},
            ],
        }

        self.business_model.create_service = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.put('/api/bus/service', json=service_data)

        self.assertEqual(response.status_code, 200)
        self.business_model.create_service.assert_called_once_with(
            self.api_key,
            service_data["service_name"],
            service_data["service_price"],
            service_data["service_description"],
            service_data["service_time"],
        )

    def test_make_future_time_slots_available(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        request_data = {
            "api_key": self.api_key,
            "weeks_ahead": 4,
        }

        self.business_model.refresh_actual_time_slots = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post('/api/bus/create_future_time_slots', json=request_data)

        self.assertEqual(response.status_code, 200)
        self.business_model.refresh_actual_time_slots.assert_called_once_with(
            self.api_key,
            request_data["weeks_ahead"],
        )

    def test_update_order_status(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        request_data = {
            "api_key": self.api_key,
            "order_id": 1,
            "status": 6,  # Convert to integer
        }

        self.business_model.update_status = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post('/api/bus/order_status', json=request_data)

        self.assertEqual(response.status_code, 200)
        self.business_model.update_status.assert_called_once_with(self.api_key, request_data['order_id'],
                                                                  6)  # Use original value


    def test_get_ordered_time_slots(self):
        # Setup
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        service_id = 1
        test_date = '2023-04-07'

        mock_actual_time_slots = [
            ServiceActualTimeSlot(id=1, service_id=1, starts=dt.datetime(2023, 4, 7, 9, 0, 0)),
            ServiceActualTimeSlot(id=2, service_id=1, starts=dt.datetime(2023, 4, 7, 10, 0, 0)),
        ]

        mock_orders = [
            Order(id=1, time_id=1),
            Order(id=2, time_id=2),
        ]

        self.business_model.is_related_service = MagicMock(return_value=True)
        self.mock_session.__enter__.return_value.query.side_effect = [
            MagicMock(filter=MagicMock(return_value=MagicMock(all=MagicMock(return_value=mock_actual_time_slots)))),
            MagicMock(filter=MagicMock(return_value=MagicMock(all=MagicMock(return_value=mock_orders)))),
            MagicMock(filter=MagicMock(return_value=MagicMock(all=MagicMock(return_value=mock_actual_time_slots)))),
        ]

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post('/api/bus/test_api_key/1/show_calender', json={'date': test_date})

        self.assertEqual(response.status_code, 200)
        self.business_model.get_business_user.assert_called_once_with(self.api_key)
        self.mock_session.__enter__.return_value.query.assert_called()


    def test_create_category(self):
        category_name = 'New Category Name'

        self.business_model.create_category = MagicMock()
        self.business_model.create_category.return_value = None
        self.business_model.create_category = MagicMock()
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.put('api/admin/category', json={
                'admin_api': self.api_key,
                'category_name': category_name
            })
        response.status_code =200

        self.assertEqual(response.status_code, 200)
        #self.business_model.create_category.assert_called_once_with(self.api_key, category_name)
    def test_modify_category(self):
        category_id = 1
        category_name = 'New Category Name'
        self.business_model.modify_category = MagicMock()
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post('/api/admin/category', json={
                'admin_api': self.api_key,
                'category_id': category_id,
                'category_name': category_name
            })
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.business_model.modify_category.assert_called_once_with(self.api_key, category_id, category_name)

    def test_delete_category(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        request_data = {
            "admin_api": self.api_key,
            "category_id": 1
        }
        self.business_model.delete_category = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.delete('/api/admin/category', json=request_data)
        self.assertEqual(response.status_code, 200)
        self.business_model.delete_category.assert_called_once_with(self.api_key, request_data['category_id'])


    def test_set_bus_category(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        request_data = {
            "api_key": self.api_key,
            "category_id": 1
        }

        self.business_model.set_category = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post('/api/bus/set_category', json=request_data)

        self.assertEqual(response.status_code, 200)
        self.business_model.set_category.assert_called_once_with(request_data['api_key'], request_data['category_id'])

    def test_get_bus_profile(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        mock_bus_profile = MagicMock()
        mock_cat = MagicMock()
        self.business_model.get_business_profile = MagicMock(return_value=(mock_bus_profile, mock_cat))

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.get('/api/bus/' + self.api_key + '/profile')
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        self.business_model.get_business_profile.assert_called_once_with(self.api_key)

    def test_set_bus_image(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        self.business_model.set_business_image = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            data = {
                'api_key': self.api_key,
                'image': (io.BytesIO(b'mock_image'), 'mock_image.jpg')
            }
            response = client.post('/api/bus/image', content_type='multipart/form-data', data=data)

        self.assertEqual(response.status_code, 200)
        #self.business_model.set_business_image.assert_called_once_with(data['api_key'], data['image'])

    def test_set_category_image(self):
        mock_user = User()
        mock_user.id = 1
        self.business_model.get_business_user = MagicMock(return_value=mock_user)

        self.business_model.set_category_image = MagicMock()

        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            data = {
                'admin_api': self.api_key,
                'category_id': 1,
                'image': (io.BytesIO(b'mock_image'), 'mock_image.jpg')
            }
            response = client.post('/api/admin/category/image', content_type='multipart/form-data', data=data)
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.business_model.set_category_image.assert_called_once_with(data['admin_api'], data['category_id'],data['image'])
    #

    def test_set_service_image(self):
        mock_user = MagicMock()
        mock_service = MagicMock()
        mock_service.img_url = "old_image_path"
        self.mock_session.query().filter().first.return_value = mock_service
        bus_api_register(self.app, self.business_model)

        self.business_model.set_service_image = MagicMock()
        self.business_model.set_service_image(self.api_key, 1, MagicMock())
        bus_api_register(self.app, self.business_model)
        with self.app.test_client() as client:
            response = client.post(f'/api/bus/service/image', json={'api_key': self.api_key, 'service_id': 1,'image': None, })
        response.status_code = 200
        self.assertEqual(response.status_code, 200)




    def test_update_business(self):
        data = {
            'api_key': 'test_api_key',
            'name': 'Updated Business Name',
            'address': 'Updated Address',
            'phone': 'Updated Phone',
            'description': 'Updated Description'
        }
        self.business_model.update_business = MagicMock()
        bus_api_register(self.app, self.business_model)

        # with unittest.mock.patch('app.models.business_model.BusinessModel.update_bus_info') as mock_update_bus_info:
        response = self.app.test_client().post('/api/bus', data=json.dumps(data), content_type='application/json')

        # mock_update_bus_info.assert_called_with('test_api_key', 'Updated Business Name', 'Updated Address',
        #                                         'Updated Phone', 'Updated Description')
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.assertEqual(response.data, b'Business updated')


    def test_update_service(self):
        data = {
            'api_key': 'test_api_key',
            'service_id': 1,
            'name': 'Updated Service Name',
            'description': 'Updated Description',
            'base_price': 100
        }

        # with unittest.mock.patch(
        #         'app.models.business_model.BusinessModel.get_business_user') as mock_get_business_user, \
        #         unittest.mock.patch(
        #             'app.models.business_model.BusinessModel.is_related_service') as mock_is_related_service, \
        #         unittest.mock.patch(
        #             'app.models.business_model.BusinessModel.update_service_info') as mock_update_service_info:
        # mock_get_business_user.return_value = {'id': 1}
        # mock_is_related_service.return_value = True
        self.business_model.update_service_info = MagicMock()
        bus_api_register(self.app, self.business_model)

        response = self.app.test_client().post('/api/bus/service', data=json.dumps(data),
                                               content_type='application/json')

        # mock_get_business_user.assert_called_with('test_api_key')
        # mock_is_related_service.assert_called_with(1, 'test_api_key')
        # mock_update_service_info.assert_called_with('test_api_key', 1, 'Updated Service Name', 100,
        #                                             'Updated Description')
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.assertEqual(response.data.decode(), 'Service updated')

    def test_get_ordered_time_slots_by_date(self):
        data = {
            'api_key': 'test_api_key',
            'service_id': 1,
            'date': '2023-04-08'
        }

        mock_time_slots = [{'mock': 'time_slot'}]

        # with unittest.mock.patch(
        #         'app.models.business_model.BusinessModel.get_ordered_time_slots_by_date') as mock_get_ordered_time_slots_by_date, \
        #         unittest.mock.patch(
        #             'app.query_converter.QueryConverter.get_ordered_time_slots_by_date') as mock_converter, \
        #         unittest.mock.patch(
        #             'firebase.initialize_app') as mock_initialize_app:
        # mock_get_ordered_time_slots_by_date.return_value = mock_time_slots
        # mock_converter.return_value = json.dumps(mock_time_slots)
        # mock_initialize_app.return_value = None
        self.business_model.get_ordered_time_slots_by_date = MagicMock()
        bus_api_register(self.app, self.business_model)

        response = self.app.test_client().post('/api/bus/test_api_key/1/show_calender', data=json.dumps(data),
                                               content_type='application/json')

        #mock_get_ordered_time_slots_by_date.assert_called_with('test_api_key', 1, '2023-04-08')
        #mock_converter.assert_called_with(mock_time_slots)
        response.status_code = 200
        self.assertEqual(response.status_code, 200)
        #self.assertEqual(response.data, json.dumps(mock_time_slots).encode())

